fix(website): layout polish, TOC padding, section reorder, and server module renames
- Fix asymmetric left/right padding across all doc pages (px-8/lg:px-12)
- Remove TOC left padding, add pt-12 top padding for breathing room
- Swap Features and CTA sections on landing page (features first)
- Reduce top padding on Features (pt-6) and CTA (pt-4) sections
- Rename content modules to .server.ts (SvelteKit convention)
- Extract shared types to types.ts
- Rename ai_traffic.go to ai-traffic.go (hyphen convention)
