+++ website/src/routes/(docs)/examples/+page.server.ts
@@ -1,5 +1,5 @@
 import type { PageServerLoad } from './$types.js';
-import { loadExamples, getCategoryLabels } from '$lib/content/examples.js';
+import { loadExamples, getCategoryLabels } from '$lib/content/examples.server.js';
 
 export const prerender = true;
